# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kitchen Pantry CRM API
  description: |
    A comprehensive RESTful API for managing food service industry relationships, interactions, and sales opportunities. 
    Built on Node.js/Express with Supabase backend integration, supporting iPad-first CRM operations.
  version: 1.0.0
  contact:
    name: Kitchen Pantry CRM Support
    email: support@kitchenpantry.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kitchenpantry.com/v1
    description: Production server
  - url: https://staging-api.kitchenpantry.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
  /organizations:
    get:
      summary: List organizations
      description: Retrieves a paginated list of organizations accessible to the authenticated user based on role and territory assignments.
      tags:
        - Organizations
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of records per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: Full-text search across organization names and descriptions
          schema:
            type: string
        - name: industry_segment
          in: query
          description: Filter by industry segment
          schema:
            type: string
            enum: [Fine Dining, Fast Food, Healthcare, Education, Corporate]
        - name: priority_level
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [A, B, C, D]
        - name: assigned_user_id
          in: query
          description: Filter by assigned user
          schema:
            type: string
            format: uuid
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, priority_level, last_interaction_date]
            default: name
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create organization
      description: Creates a new organization record with validation and business rule enforcement.
      tags:
        - Organizations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /organizations/{id}:
    get:
      summary: Get organization by ID
      description: Retrieves detailed information for a specific organization including related contacts, recent interactions, and opportunities.
      tags:
        - Organizations
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update organization
      description: Updates an existing organization with complete resource replacement.
      tags:
        - Organizations
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdateRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Partially update organization
      description: Performs partial updates to an organization, updating only specified fields.
      tags:
        - Organizations
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationPatchRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete organization
      description: Soft deletes an organization by setting is_deleted = true. Related contacts and interactions are preserved for audit purposes.
      tags:
        - Organizations
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Organization deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Supabase Auth
  schemas:
...
    Address:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "Anytown"
        state:
          type: string
          example: "CA"
        postal_code:
          type: string
          example: "90210"
        country:
          type: string
          example: "US"
      required:
        - street
        - city
        - state
        - postal_code
        - country
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 3
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true
    ResponseMetadata:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Organization schemas
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        industry_segment:
          type: string
          enum: [Fine Dining, Fast Food, Healthcare, Education, Corporate]
        priority_level:
          type: string
          enum: [A, B, C, D]
        business_type:
          type: string
        annual_revenue:
          type: number
          format: decimal
        employee_count:
          type: integer
        website_url:
          type: string
          format: uri
        primary_phone:
          type: string
        primary_email:
          type: string
          format: email
        account_status:
          type: string
          enum: [active, inactive, suspended]
        assigned_user:
          $ref: '#/components/schemas/User'
        last_interaction_date:
          type: string
          format: date-time
        next_follow_up_date:
          type: string
          format: date-time
        relationship_strength:
          type: integer
          minimum: 1
          maximum: 10
        contact_count:
          type: integer
        opportunity_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrganizationDetail:
      allOf:
        - $ref: '#/components/schemas/Organization'
        - type: object
          properties:
            billing_address:
              $ref: '#/components/schemas/Address'
            shipping_address:
              $ref: '#/components/schemas/Address'
            payment_terms:
              type: string
            credit_limit:
              type: number
              format: decimal
            notes:
              type: string
            contacts:
              type: array
              items:
                $ref: '#/components/schemas/Contact'
            recent_interactions:
              type: array
              items:
                $ref: '#/components/schemas/InteractionSummary'
            opportunities:
              type: array
              items:
                $ref: '#/components/schemas/OpportunitySummary'
    OrganizationCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        industry_segment:
          type: string
          enum: [Fine Dining, Fast Food, Healthcare, Education, Corporate]
        priority_level:
          type: string
          enum: [A, B, C, D]
        business_type:
          type: string
        annual_revenue:
          type: number
          format: decimal
          minimum: 0
        employee_count:
          type: integer
          minimum: 0
        website_url:
          type: string
          format: uri
        primary_phone:
          type: string
        primary_email:
          type: string
          format: email
        billing_address:
          $ref: '#/components/schemas/Address'
        shipping_address:
          $ref: '#/components/schemas/Address'
        payment_terms:
          type: string
        credit_limit:
          type: number
          format: decimal
          minimum: 0
        notes:
          type: string
      required:
        - name
        - industry_segment
        - priority_level
    OrganizationUpdateRequest:
      $ref: '#/components/schemas/OrganizationCreateRequest'
    OrganizationPatchRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        industry_segment:
          type: string
          enum: [Fine Dining, Fast Food, Healthcare, Education, Corporate]
        priority_level:
          type: string
          enum: [A, B, C, D]
        business_type:
          type: string
        annual_revenue:
          type: number
          format: decimal
          minimum: 0
        employee_count:
          type: integer
          minimum: 0
        website_url:
          type: string
          format: uri
        primary_phone:
          type: string
        primary_email:
          type: string
          format: email
        billing_address:
          $ref: '#/components/schemas/Address'
        shipping_address:
          $ref: '#/components/schemas/Address'
        payment_terms:
          type: string
        credit_limit:
          type: number
          format: decimal
          minimum: 0
        notes:
          type: string
    OrganizationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Organization'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'
    OrganizationDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/OrganizationDetail'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'
    OrganizationListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'

    # Contact schemas
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        full_name:
          type: string
        job_title:
          type: string
        department:
          type: string
        is_primary_contact:
          type: boolean
        is_decision_maker:
          type: boolean
        authority_level:
          type: integer
          minimum: 1
          maximum: 5
        email_primary:
          type: string
          format: email
        phone_primary:
          type: string
        mobile_phone:
          type: string
        preferred_contact_method:
          type: string
          enum: [email, phone, text, linkedin]
        preferred_contact_time:
          type: string
        linkedin_url:
          type: string
          format: uri
        communication_preferences:
          type: object
          properties:
            email_frequency:
              type: string
              enum: [daily, weekly, monthly, quarterly]
            content_types:
              type: array
              items:
                type: string
            opt_out_marketing:
              type: boolean
        last_interaction_date:
          type: string
          format: date-time
        next_follow_up_date:
          type: string
          format: date-time
        relationship_score:
          type: integer
          minimum: 1
          maximum: 10
        influence_score:
          type: integer
          minimum: 1
          maximum: 10
        engagement_level:
          type: string
          enum: [low, medium, high]
        interaction_count:
          type: integer
        organization:
          type: object
          properties:
            name:
              type: string
            industry_segment:
              type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    InteractionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        interaction_type:
          type: string
          enum: [CALL, EMAIL, MEETING, VISIT, TEXT, SOCIAL]
        interaction_date:
          type: string
          format: date-time
        subject:
          type: string
        outcome:
          type: string
          enum: [Positive, Neutral, Negative, Very Positive, Very Negative]
        user:
          type: object
          properties:
            full_name:
              type: string
...
    OpportunitySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        opportunity_name:
          type: string
        stage:
          type: string
          enum: [prospecting, qualification, needs_analysis, proposal, negotiation, closed_won, closed_lost]
        estimated_value:
          type: number
          format: decimal
        estimated_close_date:
          type: string
          format: date
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              field:
                type: string
              details:
                type: string
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
            retry_after:
              type: integer
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: UNAUTHORIZED
                message: Authentication required
                details: Valid JWT token required in Authorization header
    ForbiddenError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: FORBIDDEN
                message: Access denied
                details: Insufficient permissions for this resource
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: NOT_FOUND
                message: Resource not found
                details: The requested resource does not exist
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: BAD_REQUEST
                message: Invalid request format
                details: Request body must be valid JSON
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: CONFLICT
                message: Resource conflict
                details: Organization name already exists
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: VALIDATION_ERROR
                message: Invalid input data
                field: email_primary
                details: Email format is invalid
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: INTERNAL_ERROR
                message: Internal server error
                details: An unexpected error occurred
...